mithril_common::api_version

Trait ApiVersionDiscriminantSource

pub trait ApiVersionDiscriminantSource: Send + Sync {
    // Required method
    fn get_discriminant(&self) -> String;
}

Trait to get the discriminant that identifies the alternate openapi file to use first if found, in place of the default openapi.yml file.

Required Methods

fn get_discriminant(&self) -> String

Get the discriminant that identifies the alternate openapi file