mithril_common::crypto_helper
Module tests_setup

Test data builders for Mithril STM types, for testing purpose.

Functions

setup_certificate_chain

Instantiate a certificate chain, use this for tests only.

setup_message

Instantiate a ProtocolMessage using fake data, use this for tests only.

setup_protocol_parameters

Instantiate a ProtocolParameters, use this for tests only.

setup_signers_from_stake_distribution

Instantiate a list of protocol signers based on the given ProtocolStakeDistribution and ProtocolParameters, use this for tests only.

setup_temp_directory_for_signer

Create or retrieve a temporary directory for storing cryptographic material for a signer, use this for tests only.