mithril_common::crypto_helper

Trait KesVerifier

pub trait KesVerifier:
    Send
    + Sync
    + Debug {
    // Required method
    fn verify(
        &self,
        message: &[u8],
        signature: &Sum6KesSig,
        operational_certificate: &OpCert,
        kes_evolutions: KesEvolutions,
    ) -> StdResult<()>;
}

Trait for KES (Key Evolving Signature) verification operation.

Required Methods

fn verify( &self, message: &[u8], signature: &Sum6KesSig, operational_certificate: &OpCert, kes_evolutions: KesEvolutions, ) -> StdResult<()>

Verify the signed message and return the original message.

kes_evolutions: The KES evolutions used to verify the signature (computed from the current KES period at the time of signing minus the start KES period in the operational certificate)