mithril_common::crypto_helper

Type Alias ProtocolRegistrationError

pub type ProtocolRegistrationError = ProtocolRegistrationErrorWrapper;

Alias of a wrapper of MithrilCommon:ProtocolRegistrationErrorWrapper.

Aliased Type

pub enum ProtocolRegistrationError {
    PartyIdMissing,
    PartyIdNonExisting,
    OpCertMissing,
    OpCertInvalid,
    KesSignatureInvalid(KesEvolutions, KesPeriod),
    KesSignatureMissing,
    KesPeriodMissing,
    PoolAddressEncoding,
    CoreRegister(RegisterError),
}

Variants

PartyIdMissing

Error raised when a party id is needed but not provided

Used only for testing when SPO pool id is not certified

PartyIdNonExisting

Error raised when a party id is not available in the Cardano_stake distribution

OpCertMissing

Error raised when the operational certificate is missing

OpCertInvalid

Error raised when an operational certificate is invalid

KesSignatureInvalid(KesEvolutions, KesPeriod)

Error raised when a KES Signature verification fails

KesSignatureMissing

Error raised when a KES Signature is needed but not provided

KesPeriodMissing

Error raised when a KES Period is needed but not provided

PoolAddressEncoding

Error raised when a pool address encoding fails

CoreRegister(RegisterError)

Error raised when a core registration error occurs