mithril_common

Module entities

The entities used by, and exchanged between, the aggregator, signers and client.

Structs

AncillaryLocations

Ancillary locations of the Cardano database related files.

BlockNumber

BlockNumber is the block number of a Cardano transaction.

BlockRange

BlockRange for the Cardano chain

BlockRangesSequence

A continuous iterable sequence of block ranges.

CardanoBlock

Cardano block representation

CardanoBlockWithTransactions

Cardano block representation, including the hashes of the transactions in the block

CardanoBlocksTransactionsSigningConfig

Configuration for the signing of Cardano blocks and transactions

CardanoBlocksTransactionsSnapshot

Snapshot of a set of Cardano blocks and transactions

CardanoDatabaseSnapshot

Cardano database snapshot.

CardanoDatabaseSnapshotArtifactData

Structure holding artifacts data needed to create a Cardano database snapshot.

CardanoDbBeacon

A point in the Cardano chain at which a Mithril certificate of the Cardano Database should be produced.

CardanoStakeDistribution

Cardano Stake Distribution

CardanoTransaction

Cardano transaction representation

CardanoTransactionsSetProof

A cryptographic proof of a set of Cardano transactions is included in the global Cardano transactions set

CardanoTransactionsSigningConfig

Configuration for the signing of Cardano transactions

CardanoTransactionsSnapshot

Snapshot of a set of Cardano transactions

Certificate

Certificate represents a Mithril certificate embedding a Mithril STM multisignature

CertificateMetadata

CertificateMetadata represents the metadata associated to a Certificate

ChainPoint

The Cardano chain point which is used to identify a specific point in the Cardano chain.

ClientError

Representation of a Client Error raised by an http server

CompressionAlgorithmIter

An iterator over the variants of CompressionAlgorithm

ConfigSecret

A wrapper type for sensitive configuration values that should never be printed.

DigestsLocations

Digests locations of the Cardano database related files.

Epoch

Epoch represents a Cardano epoch

FileUri

FileUri represents a file URI used to identify the file’s location

ImmutablesLocations

Immutables locations of the Cardano database related files.

MithrilNetwork

Representation of a Mithril network

MithrilStakeDistribution

Mithril Stake Distribution

ProtocolMessage

ProtocolMessage represents a message that is signed (or verified) by the Mithril protocol

ProtocolParameters

Protocol cryptographic parameters

ServerError

Representation of a Server Error raised by a http server

SignableManifest

Stores a map of files and their hashes, with an optional signature to verify the integrity of the signed data.

SignedEntityConfig

Convert TimePoint to SignedEntityType and list allowed signed entity types and discriminants.

SignedEntityTypeDiscriminantsIter

An iterator over the variants of SignedEntityTypeDiscriminants

Signer

Signer represents a signing participant in the network

SignerWithStake

Signer represents a signing party in the network (including its stakes)

SingleSignature

SingleSignatures represent single signatures originating from a participant in the network for a digest at won lottery indexes

SlotNumber

Cardano Slot number

Snapshot

Snapshot represents a snapshot file and its metadata

StakeDistributionParty

This represents a stakeholder.

SupportedEraIter

An iterator over the variants of SupportedEra

TemplateUri

TemplateUri represents an URI pattern used to build a file’s location

TimePoint

TimePoint aggregates all types of point in the Cardano chain and is used by the state machines for their computations.

Enums

AncillaryLocation

Locations of the ancillary files.

CardanoBlockTransactionMkTreeNode

Leaf of the Merkle tree representing blocks and transactions

CardanoNetwork

The Cardano Network that is being targeted

CertificateSignature

The signature of a Certificate

CompressionAlgorithm

Compression algorithm for the snapshot archive artifacts.

DigestLocation

Locations of the immutable file digests.

EpochError

EpochError is an error triggered by an Epoch

EpochSpecifier

Represents the different ways to specify an epoch when querying the API.

ImmutablesLocation

Locations of the immutable files.

MultiFilesUri

MultiFilesUri represents a unique location uri for multiple files

ProtocolMessagePartKey

The key of a ProtocolMessage

SignedEntityType

The signed entity type that represents a type of data signed by the Mithril protocol Note: Each variant of this enum must be associated to an entry in the signed_entity_type table of the signer/aggregator nodes. The variant are identified by their discriminant (i.e. index in the enum), thus the modification of this type should only ever consist of appending new variants.

SignedEntityTypeDiscriminants

The discriminants of the SignedEntityType enum.

SingleSignatureAuthenticationStatus

Status of the authentication of the signer that emitted the signature

SupportedEra

The era that the software is running or will run

Type Aliases

BlockHash

Hash of a Cardano Block

BlockRangeLength

BlockRangeLength is the length of a block range.

CardanoEra

Cardano Era

HexEncodedAggregateVerificationKey

Hex encoded Aggregate Verification Key

HexEncodedDigest

Hex encoded Sha256 Digest

HexEncodedEraMarkersSecretKey

Hex encoded Era Markers Secret Key

HexEncodedGenesisSecretKey

Hex encoded Genesis Secret Key

HexEncodedGenesisVerificationKey

Hex encoded Genesis Verification Key

HexEncodedKey

Hex encoded key

HexEncodedKeySlice

Hex encoded key slice

HexEncodedOpCert

Hex encoded Operational Certificate

HexEncodedSingleSignature

Hex encoded Single Signature

HexEncodedVerificationKey

Hex encoded Verification Key

HexEncodedVerificationKeySignature

Hex encoded Verification Key Signature

ImmutableFileName

ImmutableFileName represents the filename, with extension, of a immutable file in the Cardano node database

ImmutableFileNumber

ImmutableFileNumber represents the id of immutable files in the Cardano node database

LotteryIndex

LotteryIndex represents the index of a Mithril single signature lottery

MagicId

Cardano Network magic identifier

PartyId

PartyId represents a signing party in Mithril protocol

ProtocolMessagePartValue

The value of a ProtocolMessage

ProtocolVersion

Protocol version

Stake

Stake represents the stakes of a participant in the Cardano chain

StakeDistribution

StakeDistribution represents the stakes of multiple participants in the Cardano chain

TotalSPOs

Number of SPOs

TransactionHash

TransactionHash is the unique identifier of a cardano transaction.