mithril_common::entities

Type Alias AncillaryFilesManifest

pub type AncillaryFilesManifest = SignableManifest<PathBuf, String>;

Alias of SignableManifest for Ancillary files

Aliased Type

struct AncillaryFilesManifest {
    pub data: BTreeMap<PathBuf, String>,
    pub signature: Option<ProtocolKey<Signature>>,
}

Fields
data: BTreeMap<PathBuf, String>

The data stored in the manifest

signature: Option<ProtocolKey<Signature>>

The signature of the manifest

Implementations

impl AncillaryFilesManifest

pub const ANCILLARY_MANIFEST_FILE_NAME: &str = "ancillary_manifest.json"

The file name used to serialize and deserialize AncillaryFilesManifest files in JSON format

pub async fn from_paths( base_directory: &Path, paths: Vec<PathBuf>, ) -> StdResult<Self>

Creates a new manifest, without signature, from the files in the provided paths

The hash of each file will be computed and stored in the manifest

pub async fn verify_data( &self, base_directory: &Path, ) -> Result<(), AncillaryFilesManifestVerifyError>

Verifies the integrity of the data in the manifest

Checks if the files in the manifest are present in the base directory and have the same hash

pub fn compute_hash(&self) -> Vec<u8>

Aggregates the hashes of all the keys and values of the manifest