Module mithril_common::messages

Messages module This module aims at providing shared structures for API communications.

Structs

CardanoTransactionSnapshotListItemMessage

Message structure of a Cardano Transactions Snapshot list item

CardanoTransactionSnapshotMessage

Message structure of a Cardano Transactions snapshot

CardanoTransactionsProofsMessage

A cryptographic proof for a set of Cardano transactions

CardanoTransactionsSetProofMessagePart

A cryptographic proof of a set of Cardano transactions is included in the global Cardano transactions set

CertificateListItemMessage

Message structure of a certificate list item

CertificateListItemMessageMetadata

CertificateListItemMessage represents the metadata associated to a CertificateListItemMessage

CertificateMessage

Message structure of a certificate

CertificateMetadataMessagePart

CertificateMetadata represents the metadata associated to a Certificate

CertificatePendingMessage

Structure to transport crate::entities::CertificatePending data.

EpochSettingsMessage

EpochSettings represents the settings of an epoch

MithrilStakeDistributionListItemMessage

Message structure of a Mithril Stake Distribution list item

MithrilStakeDistributionMessage

Message structure of a Mithril Stake Distribution

RegisterSignatureMessage

Message structure to register single signature.

RegisterSignerMessage

Register Signer Message

SignerMessagePart

Signer Message

SignerWithStakeMessagePart

Signer with Stake Message

SnapshotDownloadMessage

Message structure of a snapshot

SnapshotListItemMessage

Message structure of a snapshot list item

SnapshotMessage

Message structure of a snapshot

VerifiedCardanoTransactions

Set of transactions verified by CardanoTransactionsProofsMessage::verify.

Enums

VerifyCardanoTransactionsProofsError

Error encountered or produced by the cardano transaction proof verification.

Traits

FromMessageAdapter

From message adapter trait

ToMessageAdapter

To message adapter trait

TryFromMessageAdapter

TryFrom message adapter trait

TryToMessageAdapter

TryTo message adapter trait

Type Aliases

CardanoTransactionSnapshotListMessage

Message structure of a Cardano Transactions Snapshots list

CertificateListMessage

Message structure of a certificate list

MithrilStakeDistributionListMessage

Message structure of a Mithril Stake Distribution list

SnapshotListMessage

Message structure of a snapshot list