mithril_common::protocol
Trait ToMessage

pub trait ToMessage: Sync + Send {
    // Required method
    fn to_message(&self) -> String;
}

Trait to convert a type to a message that can be signed or verified by the Mithril protocol.

Required Methods

fn to_message(&self) -> String

Return a String representation of the message.