mithril_common
Module resource_pool

Resource pool implementation

Structs

ResourcePool

Resource pool implementation (FIFO)

ResourcePoolItem

Resource pool item which will return the resource to the pool when dropped

Enums

ResourcePoolError

ResourcePool related errors.

Traits

Reset

Reset trait which is implemented by pooled resource items. As pool resource items are mutable, this will guarantee that the pool stays consistent and that acquired resource items do not depend from previous computations.