mithril_common::resource_pool
Trait Reset

pub trait Reset {
    // Provided method
    fn reset(&mut self) -> StdResult<()> { ... }
}

Reset trait which is implemented by pooled resource items. As pool resource items are mutable, this will guarantee that the pool stays consistent and that acquired resource items do not depend from previous computations.

Provided Methods

fn reset(&mut self) -> StdResult<()>

Reset the resource