Module mithril_common::sqlite

SQLite module. This module provides a minimal yet useful Entity framework on top of SQLite with ability to perform any SQL query possible and hydrate results in Rust structs.

Structs

EntityCursor

Database query result Iterator wrapper. This wrapper allows to call entity hydration for each extracted result.

Projection

Projection is a definition of field mapping during a query. Fields come from one or several source structures (can be tables, views or sub queries) and are mapped to a Provider query as output.

ProjectionField

Each projection field is defined by

SourceAlias

Handful tool to store SQL source aliases.

WhereCondition

Where condition builder.

Enums

HydrationError

SqLite hydration error

Traits

Provider

A Provider is able to performe queries on a database and return iterator of a defined entity. It aims at being easily testable and adaptable.

SqLiteEntity

How to hydrate an entity from a SQLite result row