Trait mithril_common::sqlite::Provider

pub trait Provider<'conn> {
    type Entity;

    fn get_connection(&'conn self) -> &'conn Connection;
    fn get_projection(&self) -> &Projection;
    fn get_definition(&self, condition: Option<&str>) -> String;

    fn find(
        &'conn self,
        condition: Option<&str>,
        params: &[Value]
    ) -> Result<EntityCursor<'conn, Self::Entity>, Box<dyn Error>> { ... }
}

A Provider is able to performe queries on a database and return iterator of a defined entity. It aims at being easily testable and adaptable.

Required Associated Types

type Entity

Entity type returned by the result cursor.

Required Methods

fn get_connection(&'conn self) -> &'conn Connection

Share the connection.

fn get_projection(&self) -> &Projection

Share the projection.

fn get_definition(&self, condition: Option<&str>) -> String

Return the definition of this provider, ie the actual SQL query this provider performs.

Provided Methods

fn find(
&'conn self,
condition: Option<&str>,
params: &[Value]
) -> Result<EntityCursor<'conn, Self::Entity>, Box<dyn Error>>

Perform the parametrized definition query.