Trait mithril_common::sqlite::SqLiteEntity

pub trait SqLiteEntity {
    fn hydrate(row: Row) -> Result<Self, HydrationError>
    where
        Self: Sized;
    fn get_projection() -> Projection;
}

How to hydrate an entity from a SQLite result row

Required Methods

fn hydrate(row: Row) -> Result<Self, HydrationError>where
Self: Sized,

This method is intended to be used when creating new instances from SQL result rows. This is the place to grab data, check consistency and types and return the entity if possible.

fn get_projection() -> Projection

Construct a Projection that will allow to hydrate this SqLiteEntity.