Trait mithril_common::store::StakeStorer

pub trait StakeStorer {
    fn save_stakes<'life0, 'async_trait>(
        &'life0 self,
        epoch: Epoch,
        stakes: StakeDistribution
    ) -> Pin<Box<dyn Future<Output = Result<Option<StakeDistribution>, StoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn get_stakes<'life0, 'async_trait>(
        &'life0 self,
        epoch: Epoch
    ) -> Pin<Box<dyn Future<Output = Result<Option<StakeDistribution>, StoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn get_last_stakes<'life0, 'async_trait>(
        &'life0 self,
        last: usize
    ) -> Pin<Box<dyn Future<Output = Result<Vec<(Epoch, StakeDistribution)>, StoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Represent a way to store the stake of mithril party members.

Required Methods

fn save_stakes<'life0, 'async_trait>(
&'life0 self,
epoch: Epoch,
stakes: StakeDistribution
) -> Pin<Box<dyn Future<Output = Result<Option<StakeDistribution>, StoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Save the stakes in the store for a given epoch.

fn get_stakes<'life0, 'async_trait>(
&'life0 self,
epoch: Epoch
) -> Pin<Box<dyn Future<Output = Result<Option<StakeDistribution>, StoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Get the stakes of all party at a given epoch.

fn get_last_stakes<'life0, 'async_trait>(
&'life0 self,
last: usize
) -> Pin<Box<dyn Future<Output = Result<Vec<(Epoch, StakeDistribution)>, StoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Return the last stakes recorded in the store. This is mainly used for testing right now.