mithril_common::test

Module builder

Builders for test

Enable easier creations of complex types for test purpose

Structs

CardanoTransactionsBuilder

Builder to easily build transactions with consistent values.

CertificateChainBuilder

A builder for creating a certificate chain. For tests only.

CertificateChainBuilderContext

Context used while building a certificate chain. For tests only.

CertificateChainFixture

Fixture built from a CertificateChainBuilder, certificates are ordered from latest to genesis

MithrilFixture

A fixture of Mithril data types.

MithrilFixtureBuilder

A builder of mithril types.

SignerFixture

A signer fixture, containing a signer entity with its corresponding protocol signer and initializer

Enums

CertificateChainingMethod

Chaining method to use when building a certificate chain with the CertificateChainBuilder. For tests only.

StakeDistributionGenerationMethod

Methods that can be used to generate the stake distribution.