mithril_common::test::crypto_helper

Trait MKMapCardanoTestExtension

pub trait MKMapCardanoTestExtension<S: MKTreeStorer> {
    // Required methods
    fn from_blocks_with_transactions(
        leaves: &[CardanoBlockWithTransactions],
    ) -> StdResult<MKMap<BlockRange, MKMapNode<BlockRange, S>, S>>;
    fn compute_proof_for_transactions_hashes<H: AsRef<str>>(
        &self,
        transactions_hashes_to_prove: &[H],
        all_leaves: &[CardanoBlockWithTransactions],
    ) -> StdResult<MkSetProof<CardanoTransaction>>;
    fn compute_proof_for_blocks_hashes<H: AsRef<str>>(
        &self,
        blocks_hashes_to_prove: &[H],
        all_leaves: &[CardanoBlockWithTransactions],
    ) -> StdResult<MkSetProof<CardanoBlock>>;

    // Provided methods
    fn compute_proof_message_for_transactions_hashes<H: AsRef<str>>(
        &self,
        certificate_hash: &str,
        transactions_hashes_to_prove: &[H],
        all_leaves: &[CardanoBlockWithTransactions],
    ) -> StdResult<CardanoTransactionsProofsV2Message> { ... }
    fn compute_proof_message_for_blocks_hashes<H: AsRef<str>>(
        &self,
        certificate_hash: &str,
        blocks_hashes_to_prove: &[H],
        all_leaves: &[CardanoBlockWithTransactions],
    ) -> StdResult<CardanoBlocksProofsMessage> { ... }
}

Extension trait adding Cardano specific test utilities to MKMap

Required Methods

fn from_blocks_with_transactions( leaves: &[CardanoBlockWithTransactions], ) -> StdResult<MKMap<BlockRange, MKMapNode<BlockRange, S>, S>>

TEST ONLY - Helper to create a MKMap from a list of cardano blocks with transactions

fn compute_proof_for_transactions_hashes<H: AsRef<str>>( &self, transactions_hashes_to_prove: &[H], all_leaves: &[CardanoBlockWithTransactions], ) -> StdResult<MkSetProof<CardanoTransaction>>

TEST ONLY - Helper to create a proof for a list of transactions hashes

IMPORTANT: the all_leaves list must contains all the leaves used to create the MKMap

fn compute_proof_for_blocks_hashes<H: AsRef<str>>( &self, blocks_hashes_to_prove: &[H], all_leaves: &[CardanoBlockWithTransactions], ) -> StdResult<MkSetProof<CardanoBlock>>

TEST ONLY - Helper to create a proof for a list of blocks hashes

IMPORTANT: the all_leaves list must contains all the leaves used to create the MKMap

Provided Methods

fn compute_proof_message_for_transactions_hashes<H: AsRef<str>>( &self, certificate_hash: &str, transactions_hashes_to_prove: &[H], all_leaves: &[CardanoBlockWithTransactions], ) -> StdResult<CardanoTransactionsProofsV2Message>

TEST ONLY - Helper to create a proof message for a list of transactions hashes

IMPORTANT: the all_leaves list must contains all the leaves used to create the MKMap

fn compute_proof_message_for_blocks_hashes<H: AsRef<str>>( &self, certificate_hash: &str, blocks_hashes_to_prove: &[H], all_leaves: &[CardanoBlockWithTransactions], ) -> StdResult<CardanoBlocksProofsMessage>

TEST ONLY - Helper to create a proof message for a list of blocks hashes

IMPORTANT: the all_leaves list must contains all the leaves used to create the MKMap

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.