mithril_common::test

Module double

Test doubles

Enable unit testing with controlled inputs and predictable behavior.

Modules

fake_data

Fake data builders for testing.

fake_keys

Pre hex encoded StmTypes for testing.

Structs

DummyApiVersionDiscriminantSource

A dummy implementation of the ApiVersionDiscriminantSource trait for testing purposes.

FakeCertificaterRetriever

A fake CertificateRetriever that returns a Certificate given its hash

Traits

Dummy

A trait for giving a type a dummy value.