mithril_common::test

Module logging

Logging utilities for tests

Macros

define_test_logger

Creates a test-specific TestLogger struct that can creates preconfigured logger instances.

Structs

MemoryDrainForTest

A testing infrastructure for logging that consists of two main components:

MemoryDrainForTestInspector

A component that provides methods to analyze logs stored by MemoryDrainForTest.