Module mithril_common::test_utils

Test utilities

They contains:

A Open Api Spec tester

Some precomputed fake data and keys

A builder of MithrilFixture to generate signers alongside a stake distribution

Modules

apispec

Tools to helps validate conformity to an OpenAPI specification

fake_data

Fake data builders for testing.

fake_keys

Pre hex encoded StmTypes for testing.

test_http_server

Define a HttpServer for test that can be configured using warp filters.

Structs

MithrilFixture

A fixture of Mithril data types.

MithrilFixtureBuilder

A builder of mithril types.

SignerFixture

A signer fixture, containing a signer entity with its corresponding protocol signer and initializer

TempDir

A builder of temp directory for tests purpose.

Enums

StakeDistributionGenerationMethod

Methods that can be used to generate the stake distribution.