Crate mithril_common

Shared datatypes and traits used by Mithril rust projects

Provide:

Helpers for the Mithril STM lib with the crypto_helper.

certificate chain used to validate the Certificate Chain created by an aggregator

The entities used by, and exchanged between, the aggregator, signers and client.

Re-exports

pub use entities::CardanoNetwork;

pub use entities::MagicId;

Modules

api_version

API Version provider service

certificate_chain

Tools to retrieve, validate the Certificate Chain created by an aggregator

crypto_helper

Tools and types to abstract the use of the Mithril STM library

entities

The entities used by, and exchanged between, the aggregator, signers and client.

logging

Logging utilities for the Mithril project.

messages

Messages module This module aims at providing shared structures for API communications.

protocol

Protocol module

signable_builder

The module used for building signables

test

Test utilities

Macros

assert_dir_eq

Compare a directory against a string representing its expected structure or against another directory.

assert_equivalent

Assert that two iterators are equivalent

assert_same_json

Compare two json strings ignoring keys order

current_function

Returns the name of the function that called this macro.

current_function_path

Returns the path of the function that called this macro.

define_test_logger

Creates a test-specific TestLogger struct that can creates preconfigured logger instances.

era_deprecate

Macro used to mark the code that should be cleaned up when the new era is activated

temp_dir

Return a temporary directory based on the current function name.

temp_dir_create

Create and return a temporary directory based on the current function name.

Enums

AggregateSignatureType

The type of STM aggregate signature.

Constants

MITHRIL_AGGREGATOR_VERSION_HEADER

Mithril aggregator node version header name

MITHRIL_API_VERSION_HEADER

Mithril API protocol version header name

MITHRIL_CLIENT_TYPE_HEADER

Mithril client type of the request

MITHRIL_ORIGIN_TAG_HEADER

Mithril origin of the request

MITHRIL_SIGNER_VERSION_HEADER

Mithril signer node version header name

Type Aliases

StdError

Generic error type

StdResult

Generic result type