Crate mithril_common

Shared datatypes and traits used by Mithril rust projects

Provide:

A way to store data with the store types

Digester to compute mithril digest from a Cardano database

Helpers for the Mithril Core lib with the crypto_helper.

The entities used by, and exchanged between, the aggregator, signers and client.

useful test utilities including stubs, fake data builders, a tool validate conformity to an open api specification (apispec).

Re-exports

pub use entities::CardanoNetwork;

pub use entities::MagicId;

Modules

apispec

Tools to helps validate conformity to an OpenAPI specification

chain_observer

Tools to request metadata, like the current epoch or the stake distribution, from the Cardano

crypto_helper

Tools and types to abstract the use of the Mithril Core library

digesters

Tools to compute mithril digest from a Cardano node database.

entities

The entities used by, and exchanged between, the aggregator, signers and client.

fake_data

Fake data builders for testing.

store

Define a generic way to store data with the Store Adapters, and the StakeStorer to store stakes.

Structs

BeaconProviderImpl

A BeaconProvider using a ChainObserver and a ImmutableFileObserver.

Enums

BeaconProviderError

BeaconProvider related errors.

Constants

SIGNER_EPOCH_RECORDING_OFFSET

The epoch offset used for signers stake distribution and verification keys recording.

SIGNER_EPOCH_RETRIEVAL_OFFSET

The epoch offset used for signers stake distribution and verification keys retrieval.

Traits

BeaconProvider

Provide the current Beacon of a cardano node.