Crate mithril_common

Shared datatypes and traits used by Mithril rust projects

Provide:

A way to store data with the store types

Digester to compute mithril digest from a Cardano database

Helpers for the Mithril Core lib with the crypto_helper.

A [certificate chain] used to validate the Certificate Chain created by an aggregator

The entities used by, and exchanged between, the aggregator, signers and client.

useful test utilities including stubs, fake data builders, a tool validate conformity to an open api specification (apispec).

Modules

apispec

Tools to helps validate conformity to an OpenAPI specification

certificate_chain

Tools to retrieve, validate the Certificate Chain created by an aggregator

chain_observer

Tools to request metadata, like the current epoch or the stake distribution, from the Cardano

crypto_helper

Tools and types to abstract the use of the Mithril Core library

digesters

Tools to compute mithril digest from a Cardano node database.

entities

The entities used by, and exchanged between, the aggregator, signers and client.

fake_data

Fake data builders for testing.

store

Define a generic way to store data with the Store Adapters, and the StakeStorer to store stakes.

Structs

BeaconProviderImpl

A BeaconProvider using a ChainObserver and a ImmutableFileObserver.

Enums

BeaconProviderError

BeaconProvider related errors.

CardanoNetwork

The Cardano Network that is being targeted

Constants

MITHRIL_API_VERSION

Mithril API protocol version this is the same as the one in openapi.yml file. If you want to update this version to reflect changes in the protocol, please also update the entry in the openapi.yml

NEXT_SIGNER_EPOCH_RETRIEVAL_OFFSET

The epoch offset used to retrieve the signers stake distribution and verification keys that’s currently being signed so it can be used in the next epoch.

SIGNER_EPOCH_RECORDING_OFFSET

The epoch offset used for signers stake distribution and verification keys recording.

SIGNER_EPOCH_RETRIEVAL_OFFSET

The epoch offset used for signers stake distribution and verification keys retrieval.

Traits

BeaconProvider

Provide the current Beacon of a cardano node.

Type Definitions

MagicId

Cardano Network magic identifier