Crate mithril_common

Shared datatypes and traits used by Mithril rust projects

Provide:

A way to store data with the store types

Digester to compute mithril digest from a Cardano database

Helpers for the Mithril STM lib with the crypto_helper.

A [certificate chain] used to validate the Certificate Chain created by an aggregator

The entities used by, and exchanged between, the aggregator, signers and client.

Re-exports

pub use entities::CardanoNetwork;

pub use entities::MagicId;

Modules

api_version

API Version provider service

certificate_chain

Tools to retrieve, validate the Certificate Chain created by an aggregator

chain_observer

Tools to request metadata, like the current epoch or the stake distribution, from the Cardano

crypto_helper

Tools and types to abstract the use of the Mithril STM library

database

database module. This module contains providers and entities shared between all application types.

digesters

Tools to compute mithril digest from a Cardano node database.

entities

The entities used by, and exchanged between, the aggregator, signers and client.

era

The module used for handling eras

messages

Messages module This module aims at providing shared structures for API communications.

protocol

Protocol module

signable_builder

The module used for building signables

sqlite

SQLite module. This module provides a minimal yet useful Entity framework on top of SQLite with ability to perform any SQL query possible and hydrate results in Rust structs.

store

Define a generic way to store data with the Store Adapters, and the StakeStorer to store stakes.

test_utils

Test utilities

Macros

era_deprecate

Macro used to mark the code that should be cleaned up when the new era is activated

Structs

BeaconProviderImpl

A BeaconProvider using a ChainObserver and a ImmutableFileObserver.

Enums

BeaconProviderError

BeaconProvider related errors.

Constants

MITHRIL_API_VERSION_HEADER

Mithril API protocol version header name

MITHRIL_SIGNER_VERSION_HEADER

Mithril Signer node version header name

Traits

BeaconProvider

Provide the current Beacon of a cardano node.

Type Definitions

StdError

Generic error type

StdResult

Generic result type