mithril_common
Module cardano_block_scanner

The module used for parsing Cardano transactions

Structs

CardanoBlockScanner

Cardano block scanner

ChainReaderBlockStreamer

Block streamer that streams blocks with a Chain block reader

DumbBlockScanner

Dumb block scanner

DumbBlockStreamer

Dumb block streamer

RawCardanoPoint

Point internal representation in the Cardano chain.

ScannedBlock

A block scanned from a Cardano database

Enums

ChainScannedBlocks

ChainScannedBlocks allows to scan new blocks and handle rollbacks

Traits

BlockScanner

A scanner that can read cardano transactions in a cardano database

BlockStreamer

Trait that define how blocks are streamed from a Cardano database

BlockStreamerTestExtensions

Tests extensions methods for the BlockStreamer trait.