mithril_common
Module certificate_chain

Tools to retrieve, validate the Certificate Chain created by an aggregator

Structs

CertificateGenesisProducer

CertificateGenesisProducer is in charge of producing a Genesis Certificate

CertificateRetrieverError

CertificateRetriever related errors.

FakeCertificaterRetriever

A fake CertificateRetriever that returns a Certificate given its hash

MithrilCertificateVerifier

MithrilCertificateVerifier is an implementation of the CertificateVerifier

Enums

CertificateVerifierError

CertificateVerifier related errors.

Traits

CertificateRetriever

CertificateRetriever is in charge of retrieving a Certificate given its hash

CertificateVerifier

CertificateVerifier is the cryptographic engine in charge of verifying multi signatures and certificates