Trait mithril_common::certificate_chain::CertificateVerifier

pub trait CertificateVerifier: Send + Sync {
    // Required methods
    fn verify_multi_signature(
        &self,
        message: &[u8],
        multi_signature: &str,
        aggregate_verification_key: &str,
        protocol_parameters: &ProtocolParameters
    ) -> Result<(), CertificateVerifierError>;
    fn verify_genesis_certificate<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        certificate: &'life1 Certificate,
        genesis_verifier: &'life2 ProtocolGenesisVerifier
    ) -> Pin<Box<dyn Future<Output = Result<(), CertificateVerifierError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn verify_standard_certificate<'life0, 'life1, 'async_trait>(
        &'life0 self,
        certificate: &'life1 Certificate,
        certificate_retriever: Arc<dyn CertificateRetriever>
    ) -> Pin<Box<dyn Future<Output = Result<Option<Certificate>, CertificateVerifierError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn verify_certificate<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        certificate: &'life1 Certificate,
        certificate_retriever: Arc<dyn CertificateRetriever>,
        genesis_verifier: &'life2 ProtocolGenesisVerifier
    ) -> Pin<Box<dyn Future<Output = Result<Option<Certificate>, CertificateVerifierError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;

    // Provided method
    fn verify_certificate_chain<'life0, 'life1, 'async_trait>(
        &'life0 self,
        certificate: Certificate,
        certificate_retriever: Arc<dyn CertificateRetriever>,
        genesis_verifier: &'life1 ProtocolGenesisVerifier
    ) -> Pin<Box<dyn Future<Output = Result<(), CertificateVerifierError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
}

CertificateVerifier is the cryptographic engine in charge of verifying multi signatures and certificates

Required Methods

fn verify_multi_signature( &self, message: &[u8], multi_signature: &str, aggregate_verification_key: &str, protocol_parameters: &ProtocolParameters ) -> Result<(), CertificateVerifierError>

Verify a multi signature

fn verify_genesis_certificate<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, certificate: &'life1 Certificate, genesis_verifier: &'life2 ProtocolGenesisVerifier ) -> Pin<Box<dyn Future<Output = Result<(), CertificateVerifierError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Verify Genesis certificate

fn verify_standard_certificate<'life0, 'life1, 'async_trait>( &'life0 self, certificate: &'life1 Certificate, certificate_retriever: Arc<dyn CertificateRetriever> ) -> Pin<Box<dyn Future<Output = Result<Option<Certificate>, CertificateVerifierError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Verify Standard certificate

fn verify_certificate<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, certificate: &'life1 Certificate, certificate_retriever: Arc<dyn CertificateRetriever>, genesis_verifier: &'life2 ProtocolGenesisVerifier ) -> Pin<Box<dyn Future<Output = Result<Option<Certificate>, CertificateVerifierError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Verify if a Certificate is valid and returns the previous Certificate in the chain if exists Step 1: Check if the hash is valid (i.e. the Certificate has not been tampered by modifying its content) Step 2: Check that the multi signature is valid if it is a Standard Certificate (i.e verifiction of the Mithril multi signature) Step 3: Check that the aggregate verification key of the Certificate is registered in the previous Certificate in the chain

Provided Methods

fn verify_certificate_chain<'life0, 'life1, 'async_trait>( &'life0 self, certificate: Certificate, certificate_retriever: Arc<dyn CertificateRetriever>, genesis_verifier: &'life1 ProtocolGenesisVerifier ) -> Pin<Box<dyn Future<Output = Result<(), CertificateVerifierError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Verify that the Certificate Chain associated to a Certificate is valid TODO: see if we can borrow the certificate instead.