mithril_common
Module chain_observer

Tools to request metadata, like the current epoch or the stake distribution, from the Cardano

Structs

CardanoCliChainObserver

A ChainObserver pulling it’s data using a CardanoCliRunner.

CardanoCliRunner

A runner able to request data from a Cardano node using the Cardano Cli.

ChainObserverBuilder

Chain observer builder

FakeObserver

A Fake ChainObserver for testing purpose using fixed data.

PallasChainObserver

A runner that uses Pallas library to interact with a Cardano node using N2C Ouroboros mini-protocols

TxDatum

TxDatum represents transaction Datum.

TxDatumBuilder

TxDatumBuilder is a TxDatum builder utility.

Enums

ChainObserverError

ChainObserver related errors.

ChainObserverType

Type of chain observers available

TxDatumError

TxDatum related errors.

TxDatumFieldTypeName

Auto-generated discriminant enum variants

TxDatumFieldValue

TxDatumFieldValue represents a field value of TxDatum.

Traits

ChainObserver

Retrieve data from the cardano network

CliRunner

CliRunner trait defines the asynchronous methods for interaction with the Cardano CLI.

Type Aliases

ChainAddress

ChainAddress represents an on chain address.