Module mithril_common::chain_observer

Tools to request metadata, like the current epoch or the stake distribution, from the Cardano

Structs

CardanoCliChainObserver

A ChainObserver pulling it’s data using a CardanoCliRunner.

CardanoCliRunner

A runner able to request data from a Cardano node using the Cardano Cli.

FakeObserver

A Fake ChainObserver for testing purpose using fixed data.

Enums

ChainObserverError

ChainObserver related errors.

Traits

ChainObserver

Retrieve data from the cardano network