Module mithril_common::crypto_helper

Tools and types to abstract the use of the Mithril Core library

Structs

OpCert

Parsed Operational Certificate

ProtocolGenesisSigner

A protocol Genesis Signer that is responsible for signing the Genesis Certificate

ProtocolGenesisVerifier

A protocol Genesis Verifier that is responsible for verifying the Genesis Certificate

Enums

ProtocolGenesisError

ProtocolGenesisSigner and ProtocolGenesisVerifier related errors.

Constants

PROTOCOL_VERSION

The current protocol version

Traits

SerDeShelleyFileFormat

Trait that allows any structure that implements Serialize and DeserializeOwned to be serialized and deserialized following the Shelly json format.

Functions

key_decode_hex

Decode key from hex helper

key_encode_hex

Encode key to hex helper

Type Definitions

KESPeriod

The KES period that is used to check if the KES keys is expired

ProtocolAggregateVerificationKey

Alias of MithrilCore:StmAggrVerificationKey.

ProtocolAggregationError

Alias of MithrilCore:AggregationError.

ProtocolClerk

Alias of MithrilCore:StmClerk.

ProtocolGenesisSecretKey

Alias of Ed25519:SecretKey.

ProtocolGenesisSignature

Alias of Ed25519:Signature.

ProtocolGenesisVerificationKey

Alias of Ed25519:PublicKey.

ProtocolInitializer

Alias of a wrapper of MithrilCore:StmInitializer.

ProtocolKeyRegistration

Alias of a wrapper of MithrilCore:KeyReg.

ProtocolLotteryIndex

Alias of MithrilCore::Index.

ProtocolMultiSignature

Alias of MithrilCore:StmAggrSig.

ProtocolParameters

Alias of MithrilCore::StmParameters.

ProtocolPartyId

The id of a mithril party.

ProtocolRegistrationError

Alias of a wrapper of MithrilCore:RegisterError.

ProtocolSigner

Alias of MithrilCore:StmSigner.

ProtocolSignerVerificationKey

Alias of MithrilCore:StmVerificationKeyPoP.

ProtocolSignerVerificationKeySignature

Alias of KES:Sum6KesSig.

ProtocolSingleSignature

Alias of MithrilCore:StmSig.

ProtocolStake

Alias of MithrilCore:Stake.

ProtocolStakeDistribution

A list of Party Id associated with its Stake.

ProtocolVersion

A protocol version