Module mithril_common::crypto_helper

Tools and types to abstract the use of the Mithril STM library

Structs

OpCert

Parsed Operational Certificate

ProtocolGenesisSigner

A protocol Genesis Signer that is responsible for signing the Genesis Certificate

ProtocolGenesisVerifier

A protocol Genesis Verifier that is responsible for verifying the Genesis Certificate

Enums

ProtocolGenesisError

ProtocolGenesisSigner and ProtocolGenesisVerifier related errors.

Constants

PROTOCOL_VERSION

The current protocol version

Traits

SerDeShelleyFileFormat

Trait that allows any structure that implements Serialize and DeserializeOwned to be serialized and deserialized following the Shelly json format.

Functions

key_decode_hex

Decode key from hex helper

key_encode_hex

Encode key to hex helper

Type Definitions

KESPeriod

The KES period that is used to check if the KES keys is expired

ProtocolAggregateVerificationKey

Alias of MithrilStm:StmAggrVerificationKey.

ProtocolAggregationError

Alias of MithrilStm:AggregationError.

ProtocolClerk

Alias of MithrilStm:StmClerk.

ProtocolGenesisSecretKey

Alias of Ed25519:SecretKey.

ProtocolGenesisSignature

Alias of Ed25519:Signature.

ProtocolGenesisVerificationKey

Alias of Ed25519:PublicKey.

ProtocolInitializer

Alias of a wrapper of MithrilStm:StmInitializer.

ProtocolInitializerNotCertified

(Test only) Alias of MithrilStm:StmInitializer.

ProtocolKeyRegistration

Alias of a wrapper of MithrilStm:KeyReg.

ProtocolKeyRegistrationNotCertified

(Test only) Alias of MithrilStm:KeyReg. (Test only)

ProtocolLotteryIndex

Alias of MithrilStm::Index.

ProtocolMultiSignature

Alias of MithrilStm:StmAggrSig.

ProtocolParameters

Alias of MithrilStm::StmParameters.

ProtocolPartyId

The id of a mithril party.

ProtocolRegistrationError

Alias of a wrapper of MithrilStm:RegisterError.

ProtocolSigner

Alias of MithrilStm:StmSigner.

ProtocolSignerVerificationKey

Alias of MithrilStm:StmVerificationKeyPoP.

ProtocolSignerVerificationKeySignature

Alias of KES:Sum6KesSig.

ProtocolSingleSignature

Alias of MithrilStm:StmSig.

ProtocolStake

Alias of MithrilStm:Stake.

ProtocolStakeDistribution

A list of Party Id associated with its Stake.

ProtocolVersion

A protocol version