Trait mithril_common::crypto_helper::SerDeShelleyFileFormat

pub trait SerDeShelleyFileFormat: Serialize + DeserializeOwned {
    const TYPE: &'static str;
    const DESCRIPTION: &'static str;

    // Provided methods
    fn from_file<P: AsRef<Path>>(path: P) -> Result<Self, ParseError> { ... }
    fn to_file<P: AsRef<Path>>(&self, path: P) -> Result<(), ParseError> { ... }
}

Trait that allows any structure that implements Serialize and DeserializeOwned to be serialized and deserialized following the Shelly json format.

Required Associated Constants

const TYPE: &'static str

The type of Cardano key

const DESCRIPTION: &'static str

The description of the Cardano key

Provided Methods

fn from_file<P: AsRef<Path>>(path: P) -> Result<Self, ParseError>

Deserialize a type T: Serialize + DeserializeOwned from file following Cardano Shelley file format.

fn to_file<P: AsRef<Path>>(&self, path: P) -> Result<(), ParseError>

Serialize a type T: Serialize + DeserializeOwned to file following Cardano Shelley file format.