mithril_common

Module digesters

Tools to compute mithril digest from a Cardano node database.

Modules

cache

Cache provider to accelerate super::CardanoImmutableDigester computation by storing and reusing super::ImmutableFile digests.

Structs

CardanoImmutableDigester

A digester working directly on a Cardano DB immutables files

ComputedImmutablesDigests

Computed immutables digests

DumbImmutableDigester

A ImmutableDigester returning configurable result for testing purpose.

DumbImmutableFileObserver

An ImmutableFileObserver yielding fixed results for tests purpose.

DummyCardanoDb

A dummy cardano db.

DummyCardanoDbBuilder

A DummyCardanoDbBuilder builder.

ImmutableFile

Represent an immutable file in a Cardano node database directory

ImmutableFileSystemObserver

An ImmutableFileObserver using the filesystem.

LedgerFile

Represent an ledger file in a Cardano node database directory

Enums

ImmutableDigesterError

ImmutableDigester related Errors.

ImmutableFileCreationError

ImmutableFile::new related errors.

ImmutableFileListingError

ImmutableFile::list_completed_in_dir related errors.

ImmutableFileObserverError

ImmutableFileObserver related errors.

LedgerFileListingError

LedgerFile::list_all_in_dir related errors.

Constants

IMMUTABLE_DIR

Directory name for the immutable files.

LEDGER_DIR

Directory name for the ledger files.

VOLATILE_DIR

Directory name for the volatile files.

Traits

ImmutableDigester

A digester than can compute the digest used for mithril signatures

ImmutableFileObserver

Retrieve data on ImmutableFile from a cardano database.

Functions

immutable_trio_names

Returns the names of the files that compose an immutable trio.