Module mithril_common::digesters

Tools to compute mithril digest from a Cardano node database.

Modules

cache

Cache provider to accelerate super::CardanoImmutableDigester computation by storing and reusing super::ImmutableFile digests.

Structs

CardanoImmutableDigester

A digester working directly on a Cardano DB immutables files

DumbImmutableFileObserver

An ImmutableFileObserver yielding fixed results for tests purpose.

ImmutableFile

Represent an immutable file in a Cardano node database directory

ImmutableFileSystemObserver

An ImmutableFileObserver using the filesystem.

Enums

ImmutableDigesterError

ImmutableDigester related Errors.

ImmutableFileCreationError

ImmutableFile::new related errors.

ImmutableFileListingError

ImmutableFile::list_completed_in_dir related errors.

ImmutableFileObserverError

ImmutableFileObserver related errors.

Traits

ImmutableDigester

A digester than can compute the digest used for mithril signatures

ImmutableFileObserver

Retrieve data on ImmutableFile from a cardano database.