Module mithril_common::entities

The entities used by, and exchanged between, the aggregator, signers and client.

Structs

Beacon

Beacon represents a point in the Cardano chain at which a Mithril certificate should be produced

Certificate

Certificate represents a Mithril certificate embedding a Mithril STM multisignature

CertificateMetadata

CertificateMetadata represents the metadata associated to a Certificate

CertificatePending

CertificatePending represents a pending certificate in the process of production

Epoch

Epoch represents a Cardano epoch

InternalServerError

Representation of a Internal Server Error raised by an http server

ProtocolMessage

ProtocolMessage represents a message that is signed (or verified) by the Mithril protocol

ProtocolParameters

Protocol cryptographic parameters

Signer

Signer represents a signing participant in the network

SignerWithStake

Signer represents a signing party in the network (including its stakes)

SingleSignatures

SingleSignatures represent single signatures originating from a participant in the network for a digest at won lottery indexes

Snapshot

Snapshot represents a snapshot file and its metadata

Enums

CardanoNetwork

The Cardano Network that is being targeted

EpochError

EpochError is an error triggerred by an Epoch

ProtocolMessagePartKey

The key of a ProtocolMessage

Type Definitions

ImmutableFileNumber

ImmutableFileNumber represents the id of immutable files in the Cardano node database

LotteryIndex

LotteryIndex represents the index of a Mithril single signature lottery

MagicId

Cardano Network magic identifier

PartyId

PartyId represents a signing party in Mithril protocol

ProtocolMessagePartValue

The value of a ProtocolMessage

ProtocolVersion

Protocol version

Stake

Stake represents the stakes of a participant in the Cardano chain

StakeDistribution

StakeDistribution represents the stakes of multiple participants in the Cardano chain