mithril_common
Module era

The module used for handling eras

Modules

adapters

Module dedicated to EraReaderAdapter implementations.

Structs

EraChecker

EraChecker allows the verification of the current era

EraEpochToken

This is a response from the EraReader. It contains EraMarkers read from the adapter. It can try to cast the given markers to SupportedEras.

EraMarker

Value object that represents a tag of Era change.

EraReader

The EraReader is responsible of giving the current Era and the Era to come. It uses an EraReaderAdapter to read data from a backend.

SupportedEraIter

An iterator over the variants of SupportedEra

Enums

EraReaderError

Error type when EraReader fails to return a EraEpochToken.

SupportedEra

The era that the software is running or will run

Traits

EraReaderAdapter

Adapters are responsible of technically reading the information of EraMarkers from a backend.