mithril_common

Module messages

Messages module This module aims at providing shared structures for API communications.

Structs

AggregatorCapabilities

Capabilities of an Aggregator

AggregatorFeaturesMessage

Message advertised by an Aggregator to inform about its features

AggregatorStatusMessage

Message advertised by an aggregator to inform about its status

AncillaryMessagePart

The message part that represents the locations of the Cardano database ancillary.

CardanoBlocksTransactionsSnapshotListItemMessage

Message structure of a Cardano Transactions Snapshot list item

CardanoBlocksTransactionsSnapshotMessage

Message structure of a Cardano Blocks and Transactions snapshot

CardanoDatabaseDigestListItemMessage

Message structure of a Cardano database digest list item

CardanoDatabaseImmutableFilesRestoredMessage

Message structure of an immutable files restoration

CardanoDatabaseSnapshotListItemMessage

Message structure of a Cardano database snapshot list item

CardanoDatabaseSnapshotMessage

Cardano database snapshot.

CardanoStakeDistributionListItemMessage

Message structure of a Cardano Stake Distribution list item

CardanoStakeDistributionMessage

Message structure of a Cardano Stake Distribution

CardanoTransactionSnapshotListItemMessage

Message structure of a Cardano Transactions Snapshot list item

CardanoTransactionSnapshotMessage

Message structure of a Cardano Transactions snapshot

CardanoTransactionsProofsMessage

A cryptographic proof for a set of Cardano transactions

CardanoTransactionsProverCapabilities

Cardano transactions prover capabilities

CardanoTransactionsSetProofMessagePart

A cryptographic proof of a set of Cardano transactions is included in the global Cardano transactions set

CertificateListItemMessage

Message structure of a certificate list item

CertificateListItemMessageMetadata

CertificateListItemMessage represents the metadata associated to a CertificateListItemMessage

CertificateMessage

Message structure of a certificate

CertificateMetadataMessagePart

CertificateMetadata represents the metadata associated to a Certificate

DigestsMessagePart

The message part that represents the locations of the Cardano database digests.

EpochSettingsMessage

EpochSettings represents the settings of an epoch

ImmutablesMessagePart

The message part that represents the locations of the Cardano database immutables.

MithrilStakeDistributionListItemMessage

Message structure of a Mithril Stake Distribution list item

MithrilStakeDistributionMessage

Message structure of a Mithril Stake Distribution

ProtocolConfigurationMessage

ProtocolConfiguration represents the protocol configuration of an epoch

RegisterSignatureMessageDmq

Message structure to register single signature through the DMQ network.

RegisterSignatureMessageHttp

Message structure to register single signature through HTTP.

RegisterSignerMessage

Register Signer Message

SignerMessagePart

Signer Message

SignerWithStakeMessagePart

Signer with Stake Message

SnapshotDownloadMessage

Message structure of a snapshot

SnapshotListItemMessage

Message structure of a snapshot list item

SnapshotMessage

Message structure of a snapshot

VerifiedCardanoTransactions

Set of transactions verified by CardanoTransactionsProofsMessage::verify.

Enums

VerifyCardanoTransactionsProofsError

Error encountered or produced by the cardano transaction proof verification.

Traits

FromMessageAdapter

From message adapter trait

ToMessageAdapter

To message adapter trait

TryFromMessageAdapter

TryFrom message adapter trait

TryToMessageAdapter

TryTo message adapter trait

Type Aliases

CardanoBlocksTransactionsSnapshotListMessage

Message structure of a Cardano Transactions Snapshots list

CardanoDatabaseDigestListMessage

Message structure of a Cardano database digests list

CardanoDatabaseSnapshotListMessage

Message structure of a Cardano database snapshot list

CardanoStakeDistributionListMessage

Message structure of a Cardano Stake Distribution list

CardanoTransactionSnapshotListMessage

Message structure of a Cardano Transactions Snapshots list

CertificateListMessage

Message structure of a certificate list

MithrilStakeDistributionListMessage

Message structure of a Mithril Stake Distribution list

SnapshotListMessage

Message structure of a snapshot list