mithril_common
Module messages

Messages module This module aims at providing shared structures for API communications.

Structs

AggregatorCapabilities

Capabilities of an Aggregator

AggregatorFeaturesMessage

Message advertised by an Aggregator to inform about its features

AggregatorStatusMessage

Message advertised by an aggregator to inform about its status

CardanoDbBeaconMessagePart

A point in the Cardano chain at which a Mithril certificate of the Cardano Database should be produced.

CardanoStakeDistributionListItemMessage

Message structure of a Cardano Stake Distribution list item

CardanoStakeDistributionMessage

Message structure of a Cardano Stake Distribution

CardanoTransactionSnapshotListItemMessage

Message structure of a Cardano Transactions Snapshot list item

CardanoTransactionSnapshotMessage

Message structure of a Cardano Transactions snapshot

CardanoTransactionsProofsMessage

A cryptographic proof for a set of Cardano transactions

CardanoTransactionsProverCapabilities

Cardano transactions prover capabilities

CardanoTransactionsSetProofMessagePart

A cryptographic proof of a set of Cardano transactions is included in the global Cardano transactions set

CertificateListItemMessage

Message structure of a certificate list item

CertificateListItemMessageMetadata

CertificateListItemMessage represents the metadata associated to a CertificateListItemMessage

CertificateMessage

Message structure of a certificate

CertificateMetadataMessagePart

CertificateMetadata represents the metadata associated to a Certificate

CertificatePendingMessageDeprecated

Structure to transport crate::entities::CertificatePending data.

EpochSettingsMessage

EpochSettings represents the settings of an epoch

MithrilStakeDistributionListItemMessage

Message structure of a Mithril Stake Distribution list item

MithrilStakeDistributionMessage

Message structure of a Mithril Stake Distribution

RegisterSignatureMessage

Message structure to register single signature.

RegisterSignerMessage

Register Signer Message

SignerMessagePart

Signer Message

SignerWithStakeMessagePart

Signer with Stake Message

SnapshotDownloadMessage

Message structure of a snapshot

SnapshotListItemMessage

Message structure of a snapshot list item

SnapshotMessage

Message structure of a snapshot

VerifiedCardanoTransactions

Set of transactions verified by CardanoTransactionsProofsMessage::verify.

Enums

SignedEntityTypeMessagePart

The signed entity type that represents a type of data signed by the Mithril

VerifyCardanoTransactionsProofsError

Error encountered or produced by the cardano transaction proof verification.

Traits

FromMessageAdapter

From message adapter trait

ToMessageAdapter

To message adapter trait

TryFromMessageAdapter

TryFrom message adapter trait

TryToMessageAdapter

TryTo message adapter trait

Type Aliases

CardanoStakeDistributionListMessage

Message structure of a Cardano Stake Distribution list

CardanoTransactionSnapshotListMessage

Message structure of a Cardano Transactions Snapshots list

CertificateListMessage

Message structure of a certificate list

MithrilStakeDistributionListMessage

Message structure of a Mithril Stake Distribution list

SnapshotListMessage

Message structure of a snapshot list