Module mithril_common::messages

Messages module This module aims at providing shared structures for API communications.

Structs

CertificateListItemMessage

Message structure of a certificate list item

CertificateListItemMessageMetadata

CertificateListItemMessage represents the metadata associated to a CertificateListItemMessage

CertificateMessage

Message structure of a certificate

CertificateMetadataMessagePart

CertificateMetadata represents the metadata associated to a Certificate

CertificatePendingMessage

Structure to transport crate::entities::CertificatePending data.

EpochSettingsMessage

EpochSettings represents the settings of an epoch

MithrilStakeDistributionListItemMessage

Message structure of a Mithril Stake Distribution list item

MithrilStakeDistributionMessage

Message structure of a Mithril Stake Distribution

RegisterSignatureMessage

Message structure to register single signature.

RegisterSignerMessage

Register Signer Message

SignerMessagePart

Signer Message

SignerWithStakeMessagePart

Signer with Stake Message

SnapshotDownloadMessage

Message structure of a snapshot

SnapshotListItemMessage

Message structure of a snapshot list item

SnapshotMessage

Message structure of a snapshot

Traits

FromMessageAdapter

From message adapter trait

ToMessageAdapter

To message adapter trait

TryFromMessageAdapter

TryFrom message adapter trait

TryToMessageAdapter

TryTo message adapter trait

Type Aliases

CertificateListMessage

Message structure of a certificate list

MithrilStakeDistributionListMessage

Message structure of a Mithril Stake Distribution list

SnapshotListMessage

Message structure of a snapshot list