mithril_common
Module signable_builder

The module used for building signables

Structs

CardanoDatabaseSignableBuilder

This structure is responsible for calculating the message for incremental Cardano database.

CardanoImmutableFilesFullSignableBuilder

This structure is responsible for calculating the message for Cardano immutable files snapshots.

CardanoStakeDistributionSignableBuilder

A CardanoStakeDistributionSignableBuilder builder

CardanoTransactionsSignableBuilder

A CardanoTransactionsSignableBuilder builder

MithrilSignableBuilderService

Mithril Signable Builder Service

MithrilStakeDistributionSignableBuilder

A MithrilStakeDistributionSignableBuilder builder

SignableBuilderServiceDependencies

SignableBuilders dependencies required by the MithrilSignableBuilderService.

Traits

Artifact

Artifact is a trait for types that represent signed artifacts

Beacon

Beacon trait

BlockRangeRootRetriever

Block Range Merkle roots retriever

SignableBuilder

SignableBuilder is a trait for building a protocol message for a beacon

SignableBuilderService

ArtifactBuilder Service trait

SignableSeedBuilder

SignableSeedBuilder is a trait for building seed protocol message part values

StakeDistributionRetriever

Stake Distribution Retriever

TransactionsImporter

Cardano transactions importer