mithril_common::signable_builder

Trait SignableSeedBuilder

pub trait SignableSeedBuilder: Send + Sync {
    // Required methods
    fn compute_next_aggregate_verification_key_for_concatenation<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = StdResult<ProtocolMessagePartValue>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn compute_next_protocol_parameters<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = StdResult<ProtocolMessagePartValue>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn compute_current_epoch<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = StdResult<ProtocolMessagePartValue>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

SignableSeedBuilder is a trait for building seed protocol message part values

Required Methods

fn compute_next_aggregate_verification_key_for_concatenation<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = StdResult<ProtocolMessagePartValue>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Compute next aggregate verification key for concatenation protocol message part value

fn compute_next_protocol_parameters<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = StdResult<ProtocolMessagePartValue>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Compute next protocol parameters protocol message part value

fn compute_current_epoch<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = StdResult<ProtocolMessagePartValue>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Compute current epoch protocol message part value